pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deadline-diff-watcher"
version = "0.1.0"
description = "MVP backend for ICS deadline diff monitoring"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Deadline Diff Watcher Team" }
]
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.0,<3.0.0",
  "alembic>=1.13.0,<2.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "cryptography>=43.0.0,<46.0.0",
  "httpx>=0.27.0,<1.0.0",
  "APScheduler>=3.10.0,<4.0.0",
  "openai>=1.0.0,<2.0.0",
  "jsonschema>=4.0.0,<5.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
